move `CompiledModule`
